vim-patch:8.2.2577: compiler warning for type conversion
Problem:    Compiler warning for type conversion.
Solution:   Add a typecast. (Mike Williams)
https://github.com/vim/vim/commit/9355ae41497cbcce58ddd79f9125eb3e9dfe0a43

N/A patches for version.c:

vim-patch:8.1.0783: compiler warning for signed/unsigned

Problem:    Compiler warning for signed/unsigned.
Solution:   Add type cast.  Change type of buffer. (Ozaki Kiichi, closes vim/vim#3827)
https://github.com/vim/vim/commit/63c0ccd2b68ce854f294e6f149cc700c7f543674

Neovim was refactored to prefer char type for string functions, not char_u.

vim-patch:8.2.2152: screenpos() does not include the WinBar offset

Problem:    screenpos() does not include the WinBar offset.
Solution:   Use W_WINROW() instead of directly using w_window. (closes vim/vim#7487)
https://github.com/vim/vim/commit/8dd46e72cfb13b8de793c808ee009c45e881903a

W_WINROW() was removed so port only the test changes.
The test is currently skipped.

+++ src/nvim/screen.c
@@ -2656,7 +2656,7 @@ static int win_line(win_T *wp, linenr_T lnum, int startrow, int endrow,
           // already be in use.
           xfree(p_extra_free);
           p_extra_free = xmalloc(MAX_MCO * fdc + 1);
-          n_extra = fill_foldcolumn(p_extra_free, wp, foldinfo, lnum);
+          n_extra = (int)fill_foldcolumn(p_extra_free, wp, foldinfo, lnum);
           p_extra_free[n_extra] = NUL;
           p_extra = p_extra_free;
           c_extra = NUL;
